requirements-dev.txt
pip==22.1.2
setuptools>=34.3.3
wheel>=0.29.0
django-nose
coverage
gunicorn
flake8>=3.3.0
mccabe==0.7.0
pycodestyle>=2.3.1
pyflakes>=1.5.0
mypy>=0.501
tox
twine>=1.8.1
